UBO: Do content hashing for FloatArrays and IntArrays
Simply using `FloatArray.hashCode()` caused the issue that UBO hashing in order
to check for updates got defeated by new float arrays with the same contents being allocated.
This happened because bigvolumeviewer copies float arrays behind the scenes when assigning
them as a uniform. This commit replaces hashCode() calls on FloatArray and IntArray with
contentHashCode(), which does not take the specific reference into account, but only the
content of the array.